first version live

+++ database/migrations/2026_06_04_123614_add_mobile_number_to_users_table.php
@@ -12,13 +12,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-             $table->string('mobile', 20)->nullable()->after('password');
-        });
+        if (!Schema::hasColumn('users', 'mobile')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->string('mobile', 20)->nullable()->after('password');                
+            });
+        }
+        if (!Schema::hasColumn('pgs', 'active')) {
+            Schema::table('pgs', function (Blueprint $table) {
+                 $table->enum('active', ['0', '1'])->default('1')->after('is_verified');
+            });
+        }
         DB::statement("
             ALTER TABLE users
             MODIFY role ENUM('admin', 'student', 'owner', 'agent', 'advertise')
-            NOT NULL DEFAULT 'user'
+            NOT NULL DEFAULT 'student'
         ");
     }
 
@@ -27,13 +34,21 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('mobile');
-        });
+        if (Schema::hasColumn('users', 'mobile')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('mobile');
+            });
+        }  
+         if (Schema::hasColumn('pgs', 'active')) { 
+            Schema::table('pgs', function (Blueprint $table) {
+                $table->dropColumn('active');
+            });
+        }   
+       /*  DB::statement("UPDATE `pgs` SET `role`='student' WHERE `role`='agent'");     
+        DB::statement("UPDATE `pgs` SET `role`='student' WHERE `role`='advertise'");     
         DB::statement("
             ALTER TABLE users
             MODIFY role ENUM('admin', 'student', 'owner')
-            NOT NULL DEFAULT 'user'
-        ");
+        "); */
     }
 };
